+++ src/components/SearchBar.js
@@ -1,7 +1,7 @@
 import React, { Component } from "react";
 import { withRouter } from "react-router-dom";
 import qs from "query-string";
-import { labelAttr } from "../shared/MetadataRenderer";
+import { labelAttr } from "../lib/MetadataRenderer";
 
 class SearchBar extends Component {
   state = {
