+++ src/css/FeaturedCarousel.css
