+++ src/index.js
@@ -1,7 +1,7 @@
 import React from "react";
 import ReactDOM from "react-dom";
-import "./index.css";
-import "./colors.css";
+import "./css/index.css";
+import "./css/colors.css";
 import App from "./App";
 import * as serviceWorker from "./serviceWorker";
