+++ src/lib/MetadataRenderer.js
