+++ src/pages/collections/CollectionItemsList.js
@@ -1,7 +1,7 @@
 import React, { Component } from "react";
 import { NavLink } from "react-router-dom";
 import { Thumbnail } from "../../components/Thumbnail";
-import { arkLinkFormatted } from "../../shared/MetadataRenderer";
+import { arkLinkFormatted } from "../../lib/MetadataRenderer";
 
 import "../../css/CollectionsShowPage.css";
